Rename conflicting file for UE 5.5
