+++ Unreal/Plugins/AirSim/Source/AirSimCameraDirector.h
