+++ Unreal/Plugins/AirSim/Source/SimMode/SimModeBase.h
@@ -7,7 +7,7 @@
 #include "ParticleDefinitions.h"
 
 #include <string>
-#include "CameraDirector.h"
+#include "AirSimCameraDirector.h"
 #include "common/AirSimSettings.hpp"
 #include "common/ClockFactory.hpp"
 #include "api/ApiServerBase.hpp"
